- added detection of unreachable goals (and finding a substitute goal if none   of the fields of the original goal is reachable). - made debugging display configurable using CCL. - added caching of highlevel paths - added code that updates the pathfinder's internal description of the map   when the map changes (there are still bugs that I don't yet understand,   though). - modified the lowlevel pathfinder so that its heuristic function takes into   account information contained in the highlevel path - modified the lowlevel pathfinder so that it breaks after expanding a certain   number of nodes (it just needs to find next 2 * MAX_PATH_LENGTH steps and   with the new heuristic there's no need to compute the whole path over and   over again). - added global list of all Regions (speed optimization). - fixed some old bugs. - introduced a lot of new ones.

+++ src/pathfinder/regid.h
@@ -14,8 +14,8 @@ enum regid_usage {
 };
 
 typedef struct regid_space {
-	unsigned short RegidHigh;
-	unsigned short BitmapSize;
+	unsigned int RegidHigh;
+	unsigned int BitmapSize;
 	unsigned char *Bitmap;
 } RegidSpace;
